requirements.txt
requests==2.22.0
python-dotenv==0.12.0
python-telegram-bot==12.5.1
pysocks==1.7.1